requirements.txt
cassandra-driver==3.25.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
findspark==1.4.2
geomet==0.2.1.post1
idna==3.2
importlib-metadata==4.8.1
kafka-python==2.0.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
py4j==0.10.9
pyparsing==2.4.7
pyspark==3.1.2
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
